backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
spacy==3.7.2
nltk==3.8.1
transformers==4.35.0
torch==2.1.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0